+++ HabitTracker/presentator/src/main/java/course/doubletapp/habittracker/di/HabitListFactorySubComponent.kt
@@ -1,0 +1,14 @@
+package course.doubletapp.habittracker.di
+
+import course.doubletapp.habittracker.ui.habitlist.HabitListFragment
+import dagger.Subcomponent
+
+@Subcomponent
+interface HabitListFactorySubComponent {
+    @Subcomponent.Builder
+    interface Builder{
+        fun build(): HabitListFactorySubComponent
+    }
+
+    fun inject(habitListFragment: HabitListFragment)
+}
